# lib/api-zod/src/generated/types/adCopyVariation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdCopyVariation {
  hook: string;
  problem: string;
  solution: string;
  cta: string;
  fullCopy: string;
}