# lib/api-zod/src/generated/types/adCreativePack.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdCopyVariation } from "./adCopyVariation";
import type { ImageCreativeIdea } from "./imageCreativeIdea";
import type { OfferAndPricingAngle } from "./offerAndPricingAngle";
import type { PrimaryAngle } from "./primaryAngle";
import type { VideoAdScript } from "./videoAdScript";

export interface AdCreativePack {
  primaryAngle: PrimaryAngle;
  scrollStoppingHooks: string[];
  adCopyVariations: AdCopyVariation[];
  videoAdScript: VideoAdScript;
  imageCreativeIdeas: ImageCreativeIdea[];
  offerAndPricingAngle: OfferAndPricingAngle;
  ctaOptions: string[];
}