# lib/api-zod/src/generated/types/generateAdCreativeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateAdCreativeRequest {
  /** URL of the product image */
  imageUrl: string;
  /** Optional product category */
  productCategory?: string;
  /** Optional product name */
  productName?: string;
}