# lib/api-zod/src/generated/types/offerAndPricingAngle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OfferAndPricingAngle {
  suggestedSellingPrice: string;
  discountPrice: string;
  offerType: string;
  psychologicalReason: string;
}