# lib/api-zod/src/generated/types/videoAdScript.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoAdScript {
  /** 0-3 sec: Hook */
  hook: string;
  /** 3-8 sec: Show problem */
  problem: string;
  /** 8-15 sec: Product in action */
  productInAction: string;
  /** 15-20 sec: Result + CTA */
  resultAndCta: string;
}